Refactor Table into DataManager and Table for reuse by new Grid
I want the state management, sorting, ordering, pagination stuff here for Grid
but Grid is not a table so refactor that out into a subclass Table.
